python/requirements.txt
opencv-python>=4.8.0
face_recognition>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
pyzmq>=25.0.0
